Fix Windows cuda compilation error on floor()

+++ mmcv/ops/csrc/common/cuda/voxelization_cuda_kernel.cuh
@@ -23,20 +23,20 @@ __global__ void dynamic_voxelize_kernel(
     // To save some computation
     auto points_offset = points + index * num_features;
     auto coors_offset = coors + index * NDim;
-    int c_x = floor((points_offset[0] - coors_x_min) / voxel_x);
+    int c_x = floorf((points_offset[0] - coors_x_min) / voxel_x);
     if (c_x < 0 || c_x >= grid_x) {
       coors_offset[0] = -1;
       continue;
     }
 
-    int c_y = floor((points_offset[1] - coors_y_min) / voxel_y);
+    int c_y = floorf((points_offset[1] - coors_y_min) / voxel_y);
     if (c_y < 0 || c_y >= grid_y) {
       coors_offset[0] = -1;
       coors_offset[1] = -1;
       continue;
     }
 
-    int c_z = floor((points_offset[2] - coors_z_min) / voxel_z);
+    int c_z = floorf((points_offset[2] - coors_z_min) / voxel_z);
     if (c_z < 0 || c_z >= grid_z) {
       coors_offset[0] = -1;
       coors_offset[1] = -1;
